fix: rename local database
