+++ lib/pages/home.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../loacldb.dart';
+import '../localdb.dart';
 
 class HomePage extends StatelessWidget {
   @override
